EjeligtFelt: Har rykket gruppe funktionaliteten til Ejendom-klassen, da det er den eneste, der bruger den.
